+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for cfgrib
 ====================
 
+0.9.6.1 (unreleased)
+--------------------
+
+- Add a protocol version to the index file.
+- Fixed (for real) MULTI-FIELD messages,
+  See: `#45 <https://github.com/ecmwf/cfgrib/issues/45>`_.
+
+
 0.9.6.post1 (2019-03-07)
 ------------------------